requirements.txt
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
django-cors-headers==3.5.0
